Backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.1.8
fastapi==0.135.1
gTTS==2.5.4
groq==1.1.0
google-genai==1.66.0
pinecone==7.3.0
h11==0.16.0
idna==3.11
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.2
python-multipart==0.0.22
requests==2.32.5
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
httpx
motor==3.3.2
PyJWT==2.8.0
cryptography==42.0.5
pymongo==4.6.1